Create new `_http` module, move hacks to it
